+++ src/foundation/mod.rs
@@ -8,9 +8,11 @@
 //! It also transitively enables [`objc`](../objc/index.html).
 
 mod cmp;
+mod nsnumber;
 mod nsstring;
 
 pub use cmp::*;
+pub use nsnumber::*;
 pub use nsstring::*;
 
 #[link(name = "Foundation", kind = "framework")]